refine physical plan file structure
