+++ src/query/sql/src/executor/physical_plan_display.rs
@@ -18,40 +18,40 @@ use std::fmt::Formatter;
 use common_functions::BUILTIN_FUNCTIONS;
 use itertools::Itertools;
 
-use crate::executor::AggregateExpand;
-use crate::executor::AggregateFinal;
-use crate::executor::AggregatePartial;
-use crate::executor::AsyncSourcerPlan;
-use crate::executor::CommitSink;
-use crate::executor::CompactPartial;
-use crate::executor::ConstantTableScan;
-use crate::executor::CopyIntoTablePhysicalPlan;
-use crate::executor::CteScan;
-use crate::executor::Deduplicate;
-use crate::executor::DeletePartial;
-use crate::executor::DistributedInsertSelect;
-use crate::executor::EvalScalar;
-use crate::executor::Exchange;
-use crate::executor::ExchangeSink;
-use crate::executor::ExchangeSource;
-use crate::executor::Filter;
-use crate::executor::HashJoin;
-use crate::executor::Lambda;
-use crate::executor::Limit;
-use crate::executor::MaterializedCte;
-use crate::executor::MergeInto;
-use crate::executor::MergeIntoSource;
+use crate::executor::physical_plans::physical_aggregate_expand::AggregateExpand;
+use crate::executor::physical_plans::physical_aggregate_final::AggregateFinal;
+use crate::executor::physical_plans::physical_aggregate_partial::AggregatePartial;
+use crate::executor::physical_plans::physical_async_source::AsyncSourcerPlan;
+use crate::executor::physical_plans::physical_commit_sink::CommitSink;
+use crate::executor::physical_plans::physical_compact_partial::CompactPartial;
+use crate::executor::physical_plans::physical_constant_table_scan::ConstantTableScan;
+use crate::executor::physical_plans::physical_copy_into::CopyIntoTablePhysicalPlan;
+use crate::executor::physical_plans::physical_cte_scan::CteScan;
+use crate::executor::physical_plans::physical_deduplicate::Deduplicate;
+use crate::executor::physical_plans::physical_delete_partial::DeletePartial;
+use crate::executor::physical_plans::physical_distributed_insert_select::DistributedInsertSelect;
+use crate::executor::physical_plans::physical_eval_scalar::EvalScalar;
+use crate::executor::physical_plans::physical_exchange::Exchange;
+use crate::executor::physical_plans::physical_exchange_sink::ExchangeSink;
+use crate::executor::physical_plans::physical_exchange_source::ExchangeSource;
+use crate::executor::physical_plans::physical_filter::Filter;
+use crate::executor::physical_plans::physical_hash_join::HashJoin;
+use crate::executor::physical_plans::physical_lambda::Lambda;
+use crate::executor::physical_plans::physical_limit::Limit;
+use crate::executor::physical_plans::physical_materialized_cte::MaterializedCte;
+use crate::executor::physical_plans::physical_merge_into::MergeInto;
+use crate::executor::physical_plans::physical_merge_into::MergeIntoSource;
+use crate::executor::physical_plans::physical_project::Project;
+use crate::executor::physical_plans::physical_project_set::ProjectSet;
+use crate::executor::physical_plans::physical_range_join::RangeJoin;
+use crate::executor::physical_plans::physical_replace_into::ReplaceInto;
+use crate::executor::physical_plans::physical_row_fetch::RowFetch;
+use crate::executor::physical_plans::physical_runtime_filter_source::RuntimeFilterSource;
+use crate::executor::physical_plans::physical_sort::Sort;
+use crate::executor::physical_plans::physical_table_scan::TableScan;
+use crate::executor::physical_plans::physical_union_all::UnionAll;
+use crate::executor::physical_plans::physical_window::Window;
 use crate::executor::PhysicalPlan;
-use crate::executor::Project;
-use crate::executor::ProjectSet;
-use crate::executor::RangeJoin;
-use crate::executor::ReplaceInto;
-use crate::executor::RowFetch;
-use crate::executor::RuntimeFilterSource;
-use crate::executor::Sort;
-use crate::executor::TableScan;
-use crate::executor::UnionAll;
-use crate::executor::Window;
 use crate::plans::JoinType;
 
 impl PhysicalPlan {
