+++ src/query/sql/src/executor/physical_plans/physical_constant_table_scan.rs
@@ -1,0 +1,62 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_exception::Result;
+use common_expression::Column;
+use common_expression::DataSchemaRef;
+use common_expression::DataSchemaRefExt;
+
+use crate::executor::PhysicalPlan;
+use crate::executor::PhysicalPlanBuilder;
+use crate::ColumnSet;
+
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
+pub struct ConstantTableScan {
+    /// A unique id of operator in a `PhysicalPlan` tree.
+    /// Only used for display.
+    pub plan_id: u32,
+    pub values: Vec<Column>,
+    pub num_rows: usize,
+    pub output_schema: DataSchemaRef,
+}
+
+impl ConstantTableScan {
+    pub fn output_schema(&self) -> Result<DataSchemaRef> {
+        Ok(self.output_schema.clone())
+    }
+}
+
+impl PhysicalPlanBuilder {
+    pub(crate) async fn build_constant_table_scan(
+        &mut self,
+        scan: &crate::plans::ConstantTableScan,
+        required: ColumnSet,
+    ) -> Result<PhysicalPlan> {
+        // 1. Prune unused Columns.
+        let used: ColumnSet = required.intersection(&scan.columns).cloned().collect();
+        let (values, fields) = if used == scan.columns {
+            (scan.values.clone(), scan.schema.fields().clone())
+        } else {
+            let new_scan = scan.prune_columns(used);
+            (new_scan.values.clone(), new_scan.schema.fields().clone())
+        };
+        // 2. Build physical plan.
+        Ok(PhysicalPlan::ConstantTableScan(ConstantTableScan {
+            plan_id: self.next_plan_id(),
+            values,
+            num_rows: scan.num_rows,
+            output_schema: DataSchemaRefExt::create(fields),
+        }))
+    }
+}
