+++ src/query/sql/src/executor/physical_plans/physical_project.rs
@@ -1,0 +1,46 @@
+// Copyright 2021 Datafuse Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use common_exception::Result;
+use common_expression::DataSchemaRef;
+use common_expression::DataSchemaRefExt;
+
+use crate::executor::explain::PlanStatsInfo;
+use crate::executor::PhysicalPlan;
+use crate::optimizer::ColumnSet;
+
+#[derive(Clone, Debug, serde::Serialize, serde::Deserialize)]
+pub struct Project {
+    /// A unique id of operator in a `PhysicalPlan` tree.
+    /// Only used for display.
+    pub plan_id: u32,
+
+    pub input: Box<PhysicalPlan>,
+    pub projections: Vec<usize>,
+
+    /// Only used for display
+    pub columns: ColumnSet,
+    pub stat_info: Option<PlanStatsInfo>,
+}
+
+impl Project {
+    pub fn output_schema(&self) -> Result<DataSchemaRef> {
+        let input_schema = self.input.output_schema()?;
+        let mut fields = Vec::new();
+        for i in self.projections.iter() {
+            fields.push(input_schema.field(*i).clone());
+        }
+        Ok(DataSchemaRefExt::create(fields))
+    }
+}
